+++ test/net_http_test.rb
@@ -162,8 +162,11 @@ describe Down do
       tempfile = Down::NetHttp.download("#{$httpbin}/response-headers?Content-Disposition=inline;%20filename=\"my%2520filename.ext\"")
       assert_equal "my filename.ext", tempfile.original_filename
 
-      tempfile = Down::NetHttp.download("#{$httpbin}/response-headers?Content-Disposition=inline;%20filename=myfilename.ext%20")
-      assert_equal "myfilename.ext", tempfile.original_filename
+      tempfile = Down::NetHttp.download("#{$httpbin}/response-headers?Content-Disposition=inline;%20filename=my%2520filename.ext")
+      assert_equal "my filename.ext", tempfile.original_filename
+
+      tempfile = Down::NetHttp.download("#{$httpbin}/response-headers?Content-Disposition=inline;%20filename=\"ascii%20filename.ext\"%3B%20filename*=UTF-8''utf8%2520filename.ext")
+      assert_equal "utf8 filename.ext", tempfile.original_filename
     end
 
     it "adds #original_filename extracted from URI path if Content-Disposition is blank" do